% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIcheck.R
\name{AIcheck}
\alias{AIcheck}
\title{Error Check the Adult Index Data}
\usage{
AIcheck(streamDat, csvDir, outFile = NULL, otherTabs = NULL)
}
\arguments{
\item{streamDat}{A data frame of old and new stream mark-recapture estimates
used to estimate the lake-wide Adult Indices,
typically the output from \code{\link{AIprep}}.  The data frame must
include: \code{year},
\code{lake}, lake-stream ID \code{lscode}
(see details), population estimate
\code{PEmr}, coefficient of variation \code{CVmr}
(100% * sqrt(variance(PEmr)) / PEmr), \code{index}, a logical
identifying the index streams; \code{maintain} a logical identifying the
streams that will continue to have ongoing trapping even if not part of
the Adult Index; \code{indexContrib} a numeric, the stream population
estimate that will be used in the Adult Index (NA for new); and
\code{complete} a logical identifying streams and years for which the
Adult Index has already been estimated (FALSE for new).}

\item{csvDir}{A character scalar identifying the path where the rtf file will be
stored, e.g., \code{csvDir = "C:\\temp\\mydir"}.}

\item{outFile}{Name of the ouput rtf file, default NULL, in which case the file will be
named "YYYY Adult Index - error checking.doc" where YYYY
is the latest year represented in \code{streamDat}.}

\item{otherTabs}{A list of other tables to be printed in error check report, default NULL.
The list names will be used as captions.}
}
\value{
An error checking document as an rtf file (with the file type *.doc,
  so that MS Word will open it automatically).
}
\description{
Check the adult sea lamprey trapping data (collected for estimation of
the Adult Index) for errors.
}
\details{
Lake-stream IDs are combination of lake ID and stream ID
  e.g., 1.064 = lake ID + (stream ID)/1000.
}

