% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIestimate.R
\name{AIestimate}
\alias{AIestimate}
\title{Estimate Index of Sea Lamprey Adults}
\usage{
AIestimate(streamDat, minNMR = 2)
}
\arguments{
\item{streamDat}{A data frame of old and new stream mark-recapture estimates
used to estimate the lake-wide Adult Indices,
typically the output from \code{\link{AIprep}}.  The data frame must
include: \code{year},
\code{lake}, lake-stream ID \code{lscode}
(see details), population estimate
\code{PEmr}, coefficient of variation \code{CVmr}
(100\% * sqrt(variance(PEmr)) / PEmr), \code{index}, a logical
identifying the index streams; \code{maintain} a logical identifying the
streams that will continue to have ongoing trapping even if not part of
the Adult Index; \code{indexContrib} a numeric, the stream population
estimate that will be used in the Adult Index (NA for new);
\code{indexContribCV} a numeric, the stream CV that will be used to
generate 95\% confidence intervals for the Adult Index (NA for new); and
\code{complete} a logical identifying streams and years for which the
Adult Index has already been estimated (FALSE for new).}

\item{minNMR}{An integer scalar greater than or equal to 2,
the minimum number of mark-recapture estimates
needed in a year to generate an index, default 2.}
}
\value{
A list with 2 data frames:
  \itemize{
    \item \code{streamPE} - stream mark-recapture and Adult
  Index contributions for the incomplete rows in \code{streamDat}, with
  the same variables as \code{streamDat}; and
    \item \code{lakeIndex} - annual lake-wide Adult Indices
  with 5 columns: \code{lake}, \code{year}, the Adult Index \code{index},
  and the lower and upper 95\% confidence interval \code{ilo} and
  \code{ihi}.
  }
}
\description{
Estimate the Adult Index of sea lampreys in a single Great Lake.
}
\details{
The annual Adult Index is simply the sum of stream population estimates for
  each year.  Missing stream estimates are estimated by a lake-specific
  ANOVA model relating the log of the stream estimates to the main effects
  of each stream and each year, weighted by the inverse of the CV squared.
}
