% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIprep.R
\name{AIprep}
\alias{AIprep}
\title{Prepare the Adult Index Data}
\usage{
AIprep(csvDir, csvNew, csvOld = NULL, streamInfo = trappedStreams,
  indexStreams = lsIndex, keepStreams = lsKeep)
}
\arguments{
\item{csvDir}{A character scalar identifying the path where the csv files are
stored, e.g., \code{csvDir = "C:\\temp\\mydir"}.}

\item{csvNew}{A character scalar identifying the name of the csv file
with stream mark-recapture estimates for which Adult Indices
will be estimated (typically from the current year).  The first row of
the csv file should be column headers and must include: \code{year},
\code{lake}, lake-stream ID \code{lscode}
(see details), population estimate
\code{PEmr}, coefficient of variation \code{CVmr}
(100\% * sqrt(variance(PEmr)) / PEmr).   See details.}

\item{csvOld}{A character scalar identifying the name of the csv file
with stream mark-recapture estimates for which Adult Indices
have already been estimated (typically from previous years),
with the same variables as in \code{csvNew} plus the
previously estimated contributions \code{indexContrib} and 
\code{indexContribCV}, default NULL.
See details.}

\item{streamInfo}{A data frame with stream information to be combined with the estimates
data (csvNew and csvOld), default \code{\link{trappedStreams}}.  Set to
NULL if you don't want any stream information to be brought in.}

\item{indexStreams}{Lake-stream IDs identifying index streams, default lsIndex.}

\item{keepStreams}{Lake-stream IDs identifying streams which will continue to have ongoing
trapping even if not part of the Adult Index, default lsKeep.}
}
\value{
A single data frame that contains the information from all of the inputs,
  including new variables: \code{index}, a logical
  identifying the index streams; \code{maintain} a logical identifying the
  streams that will continue to have ongoing trapping even if not part of
  the Adult Index; \code{indexContrib} a numeric, the stream population
  estimate that will be used in the Adult Index (NA for csvNew); 
  \code{indexContribCV} a numeric, the stream CV that will be used to
  generate 95\% confidence intervals for the Adult Index (NA for new); and
  \code{complete} a logical identifying streams and years for which the
  Adult Index has already been estimated (FALSE for csvNew).
}
\description{
Read in the adult sea lamprey trapping data (collected for estimation of
the Adult Index) and prepare it for estimation.
}
\details{
Lake-stream IDs are combination of lake ID and stream ID
  e.g., 1.064 = lake ID + (stream ID)/1000.

  For both estimate files (csvNew and csvOld), there should be no missing
  values in \code{year}, \code{lake}, or \code{lscode}.
  Both files may contain variables other than those required.
  Those with the same name as variables in \code{streamInfo} will be
  replaced.  Others will be kept.
}
