% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIpresto.R
\name{AIpresto}
\alias{AIpresto}
\title{Adult Index Preparation, Error Checking, and Estimation}
\usage{
AIpresto(DIRECTORY, NEWDATARAW, STREAMDATAPREV, LAKEDATAPREV)
}
\arguments{
\item{DIRECTORY}{A character scalar identifying the path where the csv files are
stored, e.g., \code{DIRECTORY = "C:\\\\temp\\\\mydir"}.}

\item{NEWDATARAW}{A character scalar identifying the name of the csv file
with stream mark-recapture estimates for which Adult Indices
will be estimated (typically from the current year).  The first row of
the csv file should be column headers and must include: \code{year},
\code{lake}, population estimate \code{PEmr},
coefficient of variation \code{CVmr}
(100\% * sqrt(variance(PEmr)) / PEmr).}

\item{STREAMDATAPREV}{A character scalar identifying the name of the csv file
with stream mark-recapture estimates for which Adult Indices
have already been estimated (typically from previous years),
with the same variables as in \code{NEWDATARAW} plus the
previously estimated contribution \code{indexContrib}.}

\item{LAKEDATAPREV}{A character scalar identifying the name of the csv file
with annual lake-wide Adult Index estimates (typically from previous
years), with 5 columns: \code{lake}, \code{year}, \code{index}, and the
lower and upper 95\% confidence intervals \code{ilo} and \code{ihi}.}
}
\value{
Four data summaries are saved as csv files to \code{DIRECTORY}, where
  \code{YYYY} represents the most recent year of data in \code{NEWDATARAW}:
  \itemize{
    \item \code{AdultStreamYYYY.csv} - an updated version of
      \code{NEWDATARAW}
      with the additional column of the lake-stream IDs, \code{lscode},
      which are combination of lake ID and stream ID, e.g.,
      1.064 = lake ID 1 + (stream ID 64)/1000
    \item \code{AdultStreamThruYYYY.csv} - an updated version of
      \code{STREAMDATAPREV} with the latest year of data added
    \item \code{AdultLakeThruYYYY.csv} - an updated version of
      \code{LAKEDATAPREV}
      with the latest year of data added
    \item \code{AdultTargetYYYY.csv} - the calculated targets for the Adult
      Index
      of each Great Lake, with 2 columns: \code{lake} and \code{targInd}
  }

  Two rich text documents are saved as doc files (so that MS Word will open
  them automatically) to \code{DIRECTORY}:
  \itemize{
    \item \code{YYYY Adult Index - error checking.doc} - error checking
      document
    \item \code{YYYY Adult Index - draft report.doc} - draft report document
  }
}
\description{
Carry out a series of steps in the Adult Index estimation process in one
fell swoop.  This function prepares and error checks the data, estimates
the Adult Index, calculates the targets, and generates a draft report.
}
\details{
For more details on the entire process and the steps involved,
  see the vignette
  \href{https://rawgit.com/JVAdams/GLFC/master/vignettes/Adult-Index.html}{Adult Index Estimation}.
}
\examples{
\dontrun{
 library(GLFC)
 AIpresto(
   DIRECTORY = "C:\\\\TrappingData\\\\2015",
   NEWDATARAW = "TrapCatchSchaeferEstimate2015.csv",
   STREAMDATAPREV = "AdultStreamThru2014.csv",
   LAKEDATAPREV = "AdultLakeThru2014.csv")
}
}
\seealso{
\code{\link{AIprep}}, \code{\link{AIcheck}},
  \code{\link{AIestimate}}, \code{\link{AItarget}}, \code{\link{AIreport}}
}

