% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AItarget.R
\name{AItarget}
\alias{AItarget}
\title{Calculate Targets for the Adult Index}
\usage{
AItarget(lakeIndex, years = list(1994:1998, 1988:1992, 1989:1993, 1991:1995,
  1993:1997), adjust = c(1, 1, 0.25, 1, 1))
}
\arguments{
\item{lakeIndex}{A data frame of annual lake-wide Adult Indices with 4 columns:
\code{lake}, \code{year}, the Adult Index \code{index}, and
the corresponding estimates expanded out to a supposed
population estimate, \code{pe}.
The data frame may contain variables other than those required.}

\item{years}{A list of length 5 (for each Great Lake respectively: Superior, Michigan,
Huron, Erie, and Ontario), each element an integer vector of length 5
specifying the 5 years during which there were acceptable sea lamprey
wounding rates on lake trout, default
list(1994:1998, 1988:1992, 1989:1993, 1991:1995, 1993:1997).
These will be used to generate Adult Index targets.}

\item{adjust}{A numeric vector of length 5 (for each Great Lake),
specifying adjustments to be made to the calculated means to
generate Adult Index targets, default c(1, 1, 0.25, 1, 1).  Since Lake
Huron did not have a time period with acceptable sea lamprey wounding
rates, the target is set to 25\% of the mean for 1989-1993.}
}
\value{
A data frame with the calculated targets for the Adult Index and
  expanded PE of each Great Lake, with 3 columns: \code{lake},
  \code{targInd}, and \code{targPE}.
}
\description{
Calculate lake-wide targets for the Adult Index of sea lamprey
populations in the Great Lakes from the mean of specified years.
}

