% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWEFage.R
\name{DWEFage}
\alias{DWEFage}
\title{Estimate ages of sea lamprey larvae from lengths}
\usage{
DWEFage(JustLens, LenFreq = rep(1, length(JustLens)), AgeLenKey = NULL)
}
\arguments{
\item{JustLens}{A numeric vector with measured lengths of sea lamprey larvae for which
age estimates are desired. It is assumed that no young of the year (YOY),
age-0 larvae are included in the sample.}

\item{LenFreq}{A numeric vector with the number of larvae corresponding to each
measured length in \code{JustLens}. The default value will result in each
length in \code{JustLens} being counted as a single observation.}

\item{AgeLenKey}{A data frame with numeric variables \code{Length} and \code{Age} giving
measured lengths and ages of sea lamprey larvae. The default value (NULL)
will result in the use of a built in data base from 504 larvae collected
during 1993-2001.}
}
\value{
A list of length two is returned.
  The first element is a numeric matrix with the number of larvae in each
  age (row) by length (column) category. There are 7 age categories from 1
  to 7 and 31 length categories:
  <=20, >20 to <=25, ..., >160 to <=165, > 165.
  The second element is a numeric vector with the number of larvae in each
  age category.
}
\description{
Generate Estimates of larval sea lamprey ages from the measured lengths of
yearling and older larvae.
}
\examples{
DWEFage(c(120, 76, 39))
}
\references{
Haeseker, SL, ML Jones, and JR Bence. 2003.
 Estimating uncertainty in the stock-recruitment relationship for St. Marys
 River sea lampreys.
 Journal of Great Lakes Research 29(Suppl. 1):728-741.
 \href{https://doi.org/10.1016/S0380-1330(03)70527-9}{[link]}

Hoenig, JM and DM Heisey. 1987.
 Use of a log-linear model with the EM algorithm to correct estimates of
 stock composition and to convert length to age.
 Transactions of the American Fisheries Society 116(2):232-243.
}
