% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWEFerror.R
\name{DWEFerror}
\alias{DWEFerror}
\title{Error Check the Deepwater Electrofishing Data}
\usage{
DWEFerror(Dir, Catch, Lengths, Continue, Source = NULL)
}
\arguments{
\item{Dir}{A character scalar identifying the path where output files will be
stored.  Use forward slashes, e.g., \code{Dir = "C:/temp/mydir"}.}

\item{Catch}{A data frame with the catch data, typically the \code{CAT} output from
\code{\link{DWEFprep}}.}

\item{Lengths}{A data frame with the lengths data, typically the \code{LEN} output from
\code{\link{DWEFprep}}.}

\item{Continue}{A logical scalar indicating if you want to continue adding to the rtf
document after the function has run (TRUE) or if you want to end the
rtf document after the error checking (FALSE).}

\item{Source}{A named character vector with the names of the source directory and files,
c(Dir="", CatchFile="", LengthsFile="", PlotsFile=""),
typically the \code{SOURCE} output from \code{\link{DWEFprep}},
default NULL.}
}
\value{
A list with cleaned (errors removed) DWEF catch and lengths in two data
  frames (CAT2, LEN2), a character vector of the table references for any
  remaining errors (ERR), a character vector of the SOURCE directory and
  file names, and a character vector of the output file names (OUT).
}
\description{
Error check the deepwater electofishing data (including information on the
lamprey catch, the lamprey lengths, and the identification of plots that
were treated) prior to estimation.
}
\details{
If \code{Continue=FALSE}, a rich text file will be saved to the \code{DIR}
  directory with error checking text, tables, and figures.  If
  \code{Continue=TRUE}, the same rich text file will be started, but left
  open, typically to add in more text, tables, and figures generated by
  \code{\link{DWEFreport}}.
}

