% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWEFgec.R
\name{DWEFgec}
\alias{DWEFgec}
\title{Gear Efficiency Correction for Deepwater Electrofisher}
\usage{
DWEFgec(llengthmm)
}
\arguments{
\item{llengthmm}{A numeric vector of lengths (in mm) of live larval lampreys.}
}
\value{
A numeric vector equivalent to the inverse probability of capture,
  \code{p}, at length \code{llengthmm.},
  \code{logit(p) = 1.732 - 0.0229*llengthmm}.  This correction factor can
  then be multiplied by the number of lamprey of that size captured
  to estimate the catch adjusted for gear efficiency.
}
\description{
Use the relation between larval lamprey length and probability of catch
in the deepwater electrofisher to derive a gear efficiency correction factor.
}
\details{
Lake-stream IDs are combination of lake ID and stream ID
  e.g., 1.064 = lake ID + (stream ID)/1000.
  For both estimate files (csvNew and csvOld), there should be no missing
  values in \code{year}, \code{lake}, or \code{lscode}.
  Both files may contain variables other than those required.
  Those with the same name as variables in \code{streamInfo} will be
  replaced.  Others will be kept.
}
\examples{
catch <- 1:6
lenmm <- seq(80, 180, 20)
adjcatch <- catch*DWEFgec(lenmm)
cbind(catch, lenmm, adjcatch)

}
\references{
Robinson, JM, MJ Wilberg, JV Adams, and ML Jones. 2014.
 Comparing methods for estimating larval sea lamprey (Petromyzon marinus)
 density in the St. Marys River for the purposes of control.
 Journal of Great Lakes Research 40(3):739-747.
 \href{http://www.sciencedirect.com/science/article/pii/S0380133014001385}{[link]}
}

