% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWEFreport.R
\name{DWEFreport}
\alias{DWEFreport}
\title{Generate Estimates from the Deepwater Electrofishing Data}
\usage{
DWEFreport(
  Dir,
  CatchClean,
  LengthsClean,
  Plots,
  CatHist,
  LenHist,
  PlotHist,
  PEHist,
  Downstream,
  Errors,
  Outfiles,
  StratArea = SMRStratArea,
  bkill = 0.75
)
}
\arguments{
\item{Dir}{A character scalar identifying the path where output files will be
stored.  Use forward slashes, e.g., \code{Dir = "C:/temp/mydir"}.}

\item{CatchClean}{A data frame with the cleaned catch data, typically the \code{CAT2}
output from \code{\link{DWEFerror}}.}

\item{LengthsClean}{A data frame with the cleaned lengths data, including information from
larvae only (no metamorphosing juveniles), typically the \code{LEN2}
output from \code{\link{DWEFerror}}.}

\item{Plots}{A data frame with the plot data, typically the \code{PLT} output
from \code{\link{DWEFerror}}.}

\item{CatHist}{A data frame with the historic catch data, typically the \code{CAThist}
output from \code{\link{DWEFerror}}.}

\item{LenHist}{A data frame with the historic lengths data, typically the \code{LENhist}
output from \code{\link{DWEFerror}}.}

\item{PlotHist}{A data frame with the historic plot-specific estimates, typically
the \code{Plothist} output from \code{\link{DWEFerror}}.}

\item{PEHist}{A data frame with the historic whole river population estimates, typically
the \code{PEhist} output from \code{\link{DWEFerror}}.}

\item{Downstream}{Logical scalar indicating whether the downstream portion of the St. Marys
River was surveyed (TRUE) or if just the upstream portion of the river was
surveyed (FALSE).}

\item{Errors}{A character vector of table numbers corresponding to the document
produced by \code{\link{DWEFerror}}, indicating unresolved errors
remaining the DWEF data, typically the \code{ERR} output from
\code{\link{DWEFerror}}.}

\item{Outfiles}{A character vector of length seven with names for the new catch, lengths,
and plot output csv files, and the updated catch, lengths, plot, and
whole river population output csv files.  Note that updates to the whole
river population files assume post-treatment sampling only, other
situations will yield missing values for estimates.}

\item{StratArea}{Data frame with three variables: \code{inbplot} indicating whether the
stratum is in (=1) a high larval density area or not (=0), \code{region}
indicating the general location in the river (1 = North Channel,
2 = turning basin, 3 = widening part, 4 = Neebish channels, and
5 = most upstream part), and \code{haStrat} area of the stratum in
hectares.  Strata of the St. Marys River larval sea lamprey survey are
defined by \code{region} and \code{inbplot}.  By default the 2013 areas
are provided, \code{\link{SMRStratArea}}.}

\item{bkill}{Numeric scalar indicating the assumed effectiveness of Bayluscide in
treated plots, expressed as the proportion of larval sea lampreys killed,
default 0.75.}
}
\value{
Six csv files are written to \code{Dir}, with the final catch, lengths,
  and plot data for the latest year as well as the updated historic catch,
  lengths, and PE files.
}
\description{
Generate Estimates of larval sea lamprey abundance from
the deepwater electofishing data.
}
\details{
It is assumed that this function will be run immediately after the
  \code{\link{DWEFerror}} function, in which case the rtf file created by
  \code{\link{DWEFerror}} will be continued and completed by
  \code{DWEFreport}.
}
