% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/estAIndex.R
\name{estAIndex}
\alias{estAIndex}
\title{Estimate Index of Sea Lamprey Adults}
\usage{
estAIndex(indexStreams, streamPECurr, streamPEPrev = NULL, minNMR = 2,
  show = FALSE)
}
\arguments{
\item{indexStreams}{A numeric vector of IDs identifying streams to be included
in the index, e.g., 1.064 = lake code + (stream code)/1000.}

\item{streamPECurr}{A data frame of stream mark-recapture estimates from the current year,
with variables: year, lake, lscode, Emr, CVmr.}

\item{streamPEPrev}{A data frame of stream mark-recapture and Adult Index estimates
from previous years, with variables: year, lake, lscode, Emr, CVmr,
indexContrib, default NULL.}

\item{minNMR}{A numeric scalar, the minimum number of mark-recapture estimates
needed in a year to generate an index, default 2.}

\item{show}{A logical Scalar indicating if a brief summary of the results should
  by printed, default FALSE.}
}
\value{
A list with two components:
  streamPE, a data frame of stream mark-recapture and Adult Index estimates
  from previous and current years combined, with the same variables as
  \code{streamPEPrev}; and
  lakeIndex, a numeric matrix with three columns, the Adult Index,
  and the lower and upper jackknifed range.
}
\description{
Estimate the Adult Index of sea lampreys in a single Great Lake.
}
\details{
The annual Adult Index is simply the sum of the columns in \code{m} for
  each row.  The jackknifed range is produced by recalculating the index,
  leaving out one stream at a time, then scaling up the result to the same
  scale as the Adult Index based on all streams.
}
\examples{
streampe <- matrix(1:12, nrow=3, dimnames=list(1996:1998, letters[1:4]))
jackIndex(streampe)
}

