% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githubver.R
\name{githubver}
\alias{githubver}
\title{Latest GitHub Version}
\usage{
githubver(name = "JVAdams", pkg = "GLFC")
}
\arguments{
\item{name}{A character scalar identifying the name of the GitHub user where the R
package repository is located online, default \code{"JVAdams"}.}

\item{pkg}{A character scalar identifying the name of the R package,
default \code{"GLFC"}.}
}
\value{
A message is returned, reporting the version status.
}
\description{
Compare the version of the currently downloaded R package to the
latest version available on GitHub.
}
\examples{
\dontrun{
 library(GLFC)
 githubver()
}
}
