% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-data.R
\name{index2pe.Schaefer}
\alias{index2pe.Schaefer}
\title{Old Factors to Scale Up the Adult Index to a Lake-Wide Population}
\format{
A named vector of length 5 (for the 5 Great Lakes)
  with factors rounded to the nearest hundredth.
}
\source{
Great Lakes Fishery Commission (\href{http://www.glfc.org}{GLFC})
  Sea Lamprey Control Board Meeting 19-01,
  16-18 Apr 2019.
}
\description{
Lake-specific conversion factors to scale up indices of adult sea
  lamprey abundance in the Great Lakes to lake-wide population estimates,
  when indices are based on modified Schaefer estimates from 
  mark-recapture studies.
}
\author{
GLFC Trapping Task Force.
}
