% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackProp.R
\name{jackProp}
\alias{jackProp}
\title{Proportion of Contributions to Jackknifed Range that Meet Target}
\usage{
jackProp(jackRaw, target, idCols = 1:2)
}
\arguments{
\item{jackRaw}{A data frame of raw (scaled) contributions to the index.  Rows are
observations (e.g., years).  Columns include some identifiers (e.g.,
year) as well as all the individuals (e.g., locations) that were left out
when calculating the jackknifed range.}

\item{target}{A numeric scalar giving the target value with which the contributions
will be compared.}

\item{idCols}{An integer vector giving the column numbers of the identification
variables, default 1:2.  If there are no identification variables, use
\code{idCols=NULL}.}
}
\value{
A data frame with the same number of rows as \code{jackRaw} and with the
  identification variables plus a new variable, \code{pMeet} representing
  the proportion of the contributions that met the target.
}
\description{
Given a collection of raw contributions to a jackknifed range,
and a corresponding target value, this function calculates the proportion
of those contributions that are less than the target.
}
\details{
The index is simply the sum of the columns in \code{m} for
  each row.  The jackknifed range is produced by recalculating the index,
  leaving out one column at a time, then scaling up the result to the same
  scale as the index based on all columns.
}
\examples{
# 3 years of population estimates from four streams
streampe <- matrix(c(1L, 2L, 7L, 4L, 11L, 9L, 5L, 3L, 10L, 8L, 6L, 12L), nrow=3, dimnames=list(1996:1998, letters[1:4]))
jI <- jackIndex(streampe, simple=FALSE)
jraw <- jI$jack.raw
jackProp(jackRaw=jraw, target=17, idCols=NULL)
}
