% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc-data.R
\name{lsIndex}
\alias{lsIndex}
\title{Great Lakes Streams used in the Index of Adult Sea Lampreys}
\format{A list of 5 numeric vectors of stream IDs for the 5
  Great Lakes, e.g., 1.064 = (lake code) + (stream code)/1000.}
\source{
Great Lakes Fishery Commission Sea Lamprey Control Board Meeting 14-02,
  15-17 Oct 2014, Briefing Item 5 - Attachment 2,
  Transitioning to the New Adult Index in 2015
  see \href{http://www.glfc.org/boardcomm/slic/slic.php}{GLFC-SLCB}.
}
\description{
IDs identifying streams to used to generate the Adult Index.
}
\author{
GLFC Trapping Task Force.
}

