% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc-data.R
\name{lsKeep}
\alias{lsKeep}
\title{Great Lakes Streams with Commitment to Adult Sea Lamprey Trapping}
\format{A list of 5 numeric vectors of lake-stream IDs for the 5
  Great Lakes, e.g., 1.064 = (lake ID) + (stream ID)/1000.}
\source{
Great Lakes Fishery Commission (\href{http://www.glfc.org}{GLFC})
  Sea Lamprey Control Board Meeting 14-02,
  15-17 Oct 2014, Briefing Item 5 - Attachment 2,
  Transitioning to the New Adult Index in 2015.
}
\description{
IDs identifying streams which will continue to have ongoing trapping
  even if not part of the Adult Index.
}
\author{
GLFC Trapping Task Force.
}

