% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predAntilognorm.R
\name{predAntilognorm}
\alias{predAntilognorm}
\title{Unbiased Prediction of Natural Log Transformed Response on Original Scale}
\usage{
predAntilognorm(modfit, xdata, k = 0)
}
\arguments{
\item{modfit}{A fitted model object from a call to \code{\link{lm}}, \code{\link{aov}},
or \code{\link{glm}}.}

\item{xdata}{A data frame with predictor variables corresponding to those in
\code{modfit} for which predictions should be made.}

\item{k}{A numeric scalar, the constant added to the response prior to
transformation, default 0.}
}
\value{
A list with two numeric vectors, one with predicted values \code{pred}
  and one with standard deviation \code{sdpred} of those predictions,
  both on the original scale of the response.
}
\description{
Provide unbiased estimates on the original scale from a linear
model with a natural log-transformed response.
}
\examples{
fit <- aov(log(yield) ~ block + N * P + K, data=npk)
predAntilognorm(fit, npk)
}
\references{
Mood, AM, FA Graybill, DC Boes. 1974.
 Introduction to the Theory of Statistics.
 McGraw-Hill, New York.
}
