% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRpresto.R
\name{SRpresto}
\alias{SRpresto}
\title{Sea Lamprey Status Report Generation}
\usage{
SRpresto(FOLDER, INDEX.LAKE, INDEX.STREAM, MAXLARVAE,
  CONTROL = "ControlTable.csv", TROUTSUP = "TroutSuperior.csv",
  TROUTMIC = "TroutMichigan.csv", TROUTHUR = "TroutHuron.csv",
  TROUTERI = "TroutErie.csv", TROUTONT = "TroutOntario.csv",
  AXISRANGES = "AxisRanges.xls")
}
\arguments{
\item{FOLDER}{A character scalar identifying the path where all the files are
stored, e.g., \code{DIRECTORY = "C:\\\\temp\\\\mydir"}.}

\item{INDEX.LAKE}{A character scalar identifying the name of the csv file
with historic lake-wide sea lamprey adult index data, typically
\code{AdultLakeThruYYYY.csv} the output
from \code{\link{AIpresto}}.}

\item{INDEX.STREAM}{A character scalar identifying the name of the csv file
with historic stream-specific sea lamprey adult index data, typically
\code{AdultStreamThruYYYY.csv} the output
from \code{\link{AIpresto}}.}

\item{MAXLARVAE}{A character scalar identifying the name of the Excel spreadsheet
with stream-specific maximum age-1 and older sea lamprey larvae population
estimates from QAS/RS data, typically dating back to 1995.
Column headers should include: "Lake", "Stream", "Estimate", "STREAM_NAM",
"LATITUDE", and "LONGITUDE".}

\item{CONTROL}{A character scalar identifying the name of the csv file
with historic lake-wide sea lamprey control data, default
"ControlTable.csv".
The first column is the year of control.
The remaining columns are "Field Days", "TFM", and "Bayer" repeated for
each Great Lake (columns 2-4 for Superior, 5-7 Michigan, 8-10 Huron,
11-13 Erie, 14-16 Ontario) and the totals across all five lakes (columns
17-19). ??? UNITS}

\item{TROUTSUP}{A character scalar identifying the name of the csv file
with historic lake-wide lake trout data for Lake Superior
(both catch per unit effort
and sea lamprey wounding rate), default "TroutSuperior.csv".
Column headers should include: "year", "cpe", "lo", "hi", and "wound".}

\item{TROUTMIC}{A character scalar identifying the name of the csv file
with historic lake-wide lake trout data for Lake Michigan
(both catch per unit effort
and sea lamprey wounding rate), default "TroutMichigan.csv".
Column headers should include: "year", "cpe", "lo", "hi", "wound",
"wound.upper", and "wound.lower".}

\item{TROUTHUR}{A character scalar identifying the name of the csv file
with historic lake-wide lake trout data for Lake Huron
(both catch per unit effort
and sea lamprey wounding rate), default "TroutHuron.csv".
Column headers should include: "year", "cpe", "lo", "hi", and "wound".}

\item{TROUTERI}{A character scalar identifying the name of the csv file
with historic lake-wide lake trout data for Lake Erie
(both catch per unit effort
and sea lamprey wounding rate), default "TroutErie.csv".
Column headers should include: "year", "cpe", "wound",
"wound.upper", and "wound.lower".}

\item{TROUTONT}{A character scalar identifying the name of the csv file
with historic lake-wide lake trout data for Lake Ontario
(both catch per unit effort
and sea lamprey wounding rate), default "TroutOntario.csv".
Column headers should include: "year", "cpe", "wound",
"wound.upper", and "wound.lower".}

\item{AXISRANGES}{A character scalar identifying the name of the Excel spreadsheet
with specified axis ranges for the plots in the Status Report, default
"AxisRanges.xls".
Column headers should include: "order", "metric", "order", "lake", "from",
"to", "by", "from", "to", and "by" corresponding to the metric order and
name, lake order and name, then the minimum (from), maximum (to) and
spacing (by) tick marks for individual graphs (columns 5-7) and the big
one-page graphs (columns 8-10).}
}
\value{
Two data summaries are saved as csv files to \code{FOLDER}:
  \itemize{
    \item \code{StatusTargetsdd-Mon-YYYY.csv)} - Lake-wide lake trout
    wounding rate targets and sea lamprey Adult Index targets.
    \item \code{StatusMetricsdd-Mon-YYYY.csv} - Annual lake-wide
    summary of all the Status Report metrics.
  }
  One rich text documents are saved as doc files (so that MS Word will open
  them automatically) to \code{FOLDER}:
  \itemize{
    \item \code{Status Report Rough Draft dd-Mon-YYYY.doc} - Rough draft
    of the sea lamprey Status Report.
  }
}
\description{
Carry out a series of steps in the Sea Lamprey Status Report generation
process in one fell swoop.
}
\examples{
\dontrun{
 SRpresto(
  FOLDER="C:/JVA/GLFC/People/Siefkes/Status graphs/",
  INDEX.LAKE="AdultLakeThru2016.csv",
  INDEX.STREAM="AdultStreamThru2016.csv",
  MAXLARVAE="MaxLarvalEstimatesSummary2013mod2016-08-25.xls")
}
}
\seealso{
\code{\link{AIpresto}}
}
