% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRrun3.R
\name{SRrun3}
\alias{SRrun3}
\title{Three-Year Running Mean}
\usage{
SRrun3(df, x, group, time, minfreq = 2)
}
\arguments{
\item{df}{A data frame containing the data to be averaged.}

\item{x}{A character scalar, naming the element of \code{df} with the
numeric estimates to be averaged.}

\item{group}{A character scalar, naming the element of \code{df} with the
grouping variable.}

\item{time}{A character scalar, naming the element of \code{df} with the
numeric variable denoting time.}

\item{minfreq}{A numeric scalar, the minimum number of non-missing observations
required to calculate an average, default 2.}
}
\value{
A data frame with one row per group and time, giving the raw observations
  (Y) and their corresponding three-year running mean (Y.3mn).
}
\description{
Calculate the three-year moving average of a time series of data,
by the specified grouping variables.
}
\examples{
rawdat <- data.frame(
  Year = 1990 + c(0:5, 8:9, 2:3, 5:8),
  Type = rep(1:2, c(8, 6)),
  Y = c(1:12, 9, 7)*10
)
SRrun3(rawdat, "Y", "Type", "Year", minfreq=1)
SRrun3(rawdat, "Y", "Type", "Year")
SRrun3(rawdat, "Y", "Type", "Year", minfreq=3)
}
