% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRstatus.R
\name{SRstatus}
\alias{SRstatus}
\title{Determine Status and Trend}
\usage{
SRstatus(bydat, timedat, measdat, targdat, bytar = sort(unique(bydat)),
  status.length = 3, trend.length = 5, response.stat = c("Met",
  "Above"), response.trend = c("Decr.", "Steady", "Incr."))
}
\arguments{
\item{bydat}{A vector of grouping variables.}

\item{timedat}{A time variable, same length as \code{bydat}.}

\item{measdat}{A numeric vector, the estimate of interest, same length as \code{bydat}.}

\item{targdat}{A numeric vector, the target of interest, same length as \code{bytar}.}

\item{bytar}{A vector of unique grouping variables, default \code{sort(unique(bydat))}.}

\item{status.length}{A numeric scalar, the number of time units (the length of time) over
which the mean status should be assessed, default 3.  If set to NULL,
status will not be evaluated.}

\item{trend.length}{A numeric scalar, the number of time units (the length of time) over
which the trend should be assessed, default 5.  If set to NULL,
trend will not be evaluated.}

\item{response.stat}{A character vector giving the responses for estimates below and at or
above target, default c("Met", "Above")}

\item{response.trend}{A character vector giving the responses for estimates with decreasing,
steady, or increasing trends, default
c("Decr.", "Steady", "Incr.")}
}
\value{
A data frame with one row per bydat group, giving the
  time span, mean, target, and response for the status and the
  time span, slope, P value, and response for the trend.
  Column names: "bydat", "stspan", "stmean", "targdat", "status",
  "trspan", "slope", "pv", "trnd".
}
\description{
Determine the status, relative to a target, and the trend, adjusted for
autocorrelation, of a time series.
}
\details{
Significant trends in the time series are tested after automatically
  adjusting for first order autocorrelation using the iterative
  Cochrane–Orcutt procedure.  Verbeek (2004) reports that the iterated
  estimated generalized least squares (EGLS) typically performs somewhat
  better than its two-step variant in small samples (pages 100-101).
}
\examples{
rawdat <- data.frame(year=1990+c(1:6, 1:6), group=rep(1:2, c(6, 6)),
  y=c(arima.sim(n=6, list(ar=0.2)), arima.sim(n=6, list(ar=0.8))))
targetdat <- data.frame(group=1:2, targ=c(0, -1))
SRstatus(bydat=rawdat$group, timedat=rawdat$year,
  measdat=rawdat$y, targdat=targetdat$targ)
}
\references{
Verbeek M.  2004.  A Guide to Modern Econometrics.
    John Wiley & Sons Ltd, West Sussex.
}
\seealso{
\code{\link{cochrane.orcutt.jva}}
}
