% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRtrend5.R
\name{SRtrend5}
\alias{SRtrend5}
\title{Five-Year Linear Trend}
\usage{
SRtrend5(df, x, group, time, lasttime = NULL)
}
\arguments{
\item{df}{A data frame containing the data to be averaged.}

\item{x}{A character scalar, naming the element of \code{df} with the
numeric estimates to be averaged.}

\item{group}{A character scalar, naming the element of \code{df} with the
grouping variable.}

\item{time}{A character scalar, naming the element of \code{df} with the
numeric variable denoting time.}

\item{lasttime}{A numeric scalar, the last value of time that should be used to calculate
the trend. The default, NULL, results in \code{lasttime} being set to the
maximum time in \code{df}.}
}
\value{
A data frame with one row per group and last time, giving the intercept
  (Y.5nt), slope (Y.5sl), and P value (Y.5pv) for the five-year linear trend.
}
\description{
Calculate the five-year linear trend of a time series of data,
by the specified grouping variables.
}
\examples{
rawdat <- data.frame(
  Year = 1990 + c(0:5, 8:9, 2:3, 5:8),
  Type = rep(1:2, c(8, 6)),
  Y = c(1:12, 9, 7)*10
)
SRtrend5(rawdat, "Y", "Type", "Year")
SRtrend5(rawdat, "Y", "Type", "Year", lasttime=1995)
}
