% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc-data.R
\name{trappedStreams}
\alias{trappedStreams}
\title{General Information on Great Lakes Streams Trapped for Adult Sea Lampreys}
\format{A data frame with 8 elements: lake (lake code),
  lscode (lake + strcode/1000), country, strcode (stream code),
  estr (stream ID for Empiric Stream Treatment Ranking),
  strname (stream name), lat (latitude), long (longitude).}
\source{
Great Lakes Fishery Commission spawner model data base,
  last updated 12 May 2015.
}
\description{
Location information on trapped streams (past and present).
}
\author{
GLFC Trapping Task Force.
}

