% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AIreport.R
\name{AIreport}
\alias{AIreport}
\title{Adult Index Report}
\usage{
AIreport(streamPEs, lakeIPEs, targets, csvDir, outFile = NULL,
  proptargets = NULL)
}
\arguments{
\item{streamPEs}{A data frame of "complete" stream mark-recapture estimates
(meaning all contributions to the Adult Indices have
already been calculated).  The data frame must
include: \code{year},
\code{lake}, lake-stream ID \code{lscode}
(see details), population estimate
\code{PEmr}, coefficient of variation \code{CVmr}
(100% * sqrt(variance(PEmr)) / PEmr), \code{index}, a logical
identifying the index streams; \code{maintain} a logical identifying the
streams that will continue to have ongoing trapping even if not part of
the Adult Index; \code{indexContrib} a numeric, the stream population
estimate that will be used in the Adult Index (NA for new); and
\code{complete} a logical identifying streams and years for which the
Adult Index has already been estimated (should be all TRUE).}

\item{lakeIPEs}{A data frame of annual lake-wide Adult Indices with 8 columns:
\code{lake}, \code{year}, the Adult Index \code{index}, its associated
lower and upper jackknifed range \code{jlo} and \code{jhi},
and the corresponding expansion to a supposed population estimate,
\code{pe}, \code{pelo} and \code{pehi}.
The data frame may contain variables other than those required.}

\item{targets}{A data frame with the calculated targets for the Adult Index and
expanded PE of each Great Lake, with 5 rows (Superior, Michigan, Huron,
Erie, Ontario) and 3 columns: \code{lake},
\code{targInd}, and \code{targPE}, typically the output from
\code{\link{AItarget}}.}

\item{csvDir}{A character scalar identifying the path where the rtf file will be
stored, e.g., \code{csvDir = "C:\\temp\\mydir"}.}

\item{outFile}{Name of the ouput rtf file, default NULL, in which case the file will be
named "YYYY Adult Index - draft report.doc" where YYYY
is the latest year represented in \code{streamDat}.}

\item{proptargets}{A data frame with any proposed targets for the Adult Index,
with 2 columns \code{lake} and \code{targInd}, default NULL.
May have from zero to several rows for a single Great Lake.}
}
\value{
A draft report document as an rtf file (with the file type *.doc,
  so that MS Word will open it automatically).
}
\description{
Create a draft template-style report of the Adult Index estimates of
sea lamprey in the Great Lakes.
}
\details{
Lake-stream IDs are combination of lake ID and stream ID
  e.g., 1.064 = lake ID + (stream ID)/1000.
}

