% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DWEFprep.R
\name{DWEFprep}
\alias{DWEFprep}
\title{Prepare the Deepwater Electrofishing Data}
\usage{
DWEFprep(Dir, CatchFile, LengthsFile, PlotsFile, TRTtiming = "AFTER",
  b4plots = NULL)
}
\arguments{
\item{Dir}{A character scalar identifying the path where the data files are
stored.  Use forward slashes, e.g., \code{Dir = "C:/temp/mydir"}.}

\item{CatchFile}{A character scalar identifying the name of the *.xl* or *.dbf file with
catch data.  The file should have at least the following 19 columns,
named in the header row:
SAMPID, LATITUDE, LONGITUDE, STIME, BOAT, SAMPLE, DEPTH, SUB_MAJOR,
SUB_MINOR1, SUB_MINOR2, GPSDATE, HAB_TYPE, SL_TOTAL, AB_TOTAL, I_TOTAL,
COMMENT, NEW_NUMB, INBPLOT, REGION
(these last 3 columns are added in using ArcInfo).
See details.}

\item{LengthsFile}{A character vector identifying the names of the *.xl* or *.dbf files with
the lengths data.  The files should have at least the following 2 columns,
named in the header row: SAMPID, LENGTH.
See details.}

\item{PlotsFile}{A character vector identifying the name of the *.xl* or *.dbf file with
the treatment plots data.  The files should have at least the following 3
columns, named in the header row:
AREA, Plot_09, Treat_YYYY, where YYYY is the current year.
Treat_YYYY is equal to 1 if the plot was treated that year, equal to 0
otherwise.  If a plot was treated twice in one year, it will be listed
on two separate rows, each with Treat_YYYY=1.
See details.}

\item{TRTtiming}{A character scalar identifying the timing of the assessment survey relative
to treatment.  "AFTER" if all plots were surveyed AFTER they were treated
(the default), "BEFORE" if all plots were survey BEFORE they were treated,
"NONE" if no plots were treated, and "MIXED" if some plots were surveyed
before and some plots were surveyed after treatment.}

\item{b4plots}{A numeric vector identifying the plots that were surveyed BEFORE they
were treated.  A value for this should only be provided if
\code{TRTtiming} is set to "MIXED" (default NULL).}
}
\value{
A list with DWEF catch, length, and plot data in three data frames (CAT,
  LEN, PLT) and a character vector of the SOURCE directory and file names.
  The plot data is reorganized to have only one row per plot, with the
  trtd variable indicating the number of treatments each plot received
  that year.
}
\description{
Read in the deepwater electofishing data (including information on the
lamprey catch, the lamprey lengths, and the identification of plots that
were treated) and prepare them for estimation.
}
\details{
The order of the columns and the case of the column names in the
  \code{CatchFile}, \code{LengthsFile}, \code{PlotsFile} files are
  unimportant.
  Additional columns may be also be part of theses files,
  but they will be ignored.
}

