% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-data.R
\name{index2pe}
\alias{index2pe}
\title{Factors to Scale Up the Adult Index to a Lake-Wide Population}
\format{A named vector of length 5 (for the 5 Great Lakes)
  with factors rounded to the nearest hundredth.}
\source{
Great Lakes Fishery Commission (\href{http://www.glfc.org}{GLFC})
  Sea Lamprey Control Board Meeting 14-02,
  15-17 Oct 2014, Briefing Item 5 - Attachment 2,
  Transitioning to the New Adult Index in 2015.
}
\description{
Lake-specific conversion factors to scale up indices of adult sea
  lamprey abundance in the Great Lakes to lake-wide population estimates.
}
\author{
GLFC Trapping Task Force.
}
