% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackIndex.R
\name{jackIndex}
\alias{jackIndex}
\title{Index with Jackknifed Range}
\usage{
jackIndex(m)
}
\arguments{
\item{m}{A numeric matrix of estimates contributing to the index.  Rows are
observations (e.g., years).  Columns are individuals (e.g., locations).}
}
\value{
A numeric matrix with 3 columns, the index, and the lower and
  upper jackknifed range of the index.
}
\description{
Given a collection of estimates contributing to an index, this function
provides the estimated index (sum) of observations (across a row) and
the jackknifed range of the index, the result of leaving out
an entire column (e.g., a location) one at a time.
}
\details{
The index is simply the sum of the columns in \code{m} for
  each row.  The jackknifed range is produced by recalculating the index,
  leaving out one column at a time, then scaling up the result to the same
  scale as the index based on all columns.
}
\examples{
# 3 years of population estimates from four streams
streampe <- matrix(1:12, nrow=3, dimnames=list(1996:1998, letters[1:4]))
jackIndex(streampe)
}

