% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jackIndex.R
\name{jackIndex}
\alias{jackIndex}
\title{Index of Sea Lamprey Adults with Jackknifed Range}
\usage{
jackIndex(m)
}
\arguments{
\item{m}{A numeric matrix of stream run size estimates with
observation years as rows and individual streams as columns.}
}
\value{
A numeric matrix with 3 columns, the Adult Index, and the lower and
  upper jackknifed range.
}
\description{
Estimated Adult Index of sea lamprey with the observed range
in the index when one stream
at a time is excluded from the estimation.
}
\details{
The annual Adult Index is simply the sum of the columns in \code{m} for
  each row.  The jackknifed range is produced by recalculating the index,
  leaving out one stream at a time, then scaling up the result to the same
  scale as the Adult Index based on all streams.
}
\examples{
streampe <- matrix(1:12, nrow=3, dimnames=list(1996:1998, letters[1:4]))
jackIndex(streampe)
}

