% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRstatus.R
\name{SRstatus}
\alias{SRstatus}
\title{Determining Status and Trends}
\usage{
SRstatus(
  bydat,
  timedat,
  measdat,
  targdat,
  bytar = sort(unique(bydat)),
  status.length = 3,
  trend.length = 5
)
}
\arguments{
\item{bydat}{A vector of grouping variables.}

\item{timedat}{A time variable, same length as \code{bydat}.}

\item{measdat}{A numeric vector, the estimate of interest, same length as \code{bydat}.}

\item{targdat}{A numeric vector, the target of interest, same length as \code{bytar}.}

\item{bytar}{A vector of unique grouping variables, default \code{sort(unique(bydat))}.}

\item{status.length}{A numeric scalar, the number of time units (the length of time) over
which the mean status should be assessed, default 3.  If set to NULL,
status will not be evaluated.}

\item{trend.length}{A numeric scalar, the number of time units (the length of time) over
which the trend should be assessed, default 5.  If set to NULL,
trend will not be evaluated.}
}
\value{
A data frame with one row per bydat group, giving the
  time span, mean, target, and response for the status and the
  time span, slope, P value, and response for the trend.
  Column names: "bydat", "stspan", "stmean", "targdat", "status",
  "trspan", "slope", "pv", "trnd".
}
\description{
Determine the status and trends of a time series relative to a target.
}
\examples{
 rawdat <- data.frame(year=1990+c(1:6, 1:6), group=rep(1:2, c(6, 6)),
  x=c(1:10, 9, 7)*10)
 targetdat <- data.frame(group=1:2, targ=c(30, 140))
 SRstatus(bydat=rawdat$group, timedat=rawdat$year,
  measdat=rawdat$x, targdat=targetdat$targ)
}
