% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-data.R
\name{trappedStreams}
\alias{trappedStreams}
\title{General Information on Great Lakes Streams Trapped for Adult Sea Lampreys}
\format{
A data frame with 8 elements: \code{lake} (lake ID),
  \code{lscode} (lake-stream ID, lake + strcode/1000), \code{country}, 
  \code{strcode} (stream ID),
  \code{estr} (stream ID for Empiric Stream Treatment Ranking),
  \code{strname} (stream name), \code{lat} (latitude), 
  \code{long} (longitude).
}
\source{
Great Lakes Fishery Commission (\href{http://www.glfc.org}{GLFC})
  spawner model data base, last updated 12 May 2015.
}
\description{
Location information on trapped streams (past and present).
}
\author{
GLFC Trapping Task Force.
}
